LPD-39639 Create two separate change tracking playwright test folders
