+++ modules/test/playwright/tests/change-tracking-web-2/test.properties
@@ -1,0 +1,5 @@
+##
+## Testray
+##
+
+    testray.main.component.name=Publications
