+++ test.properties
@@ -8077,7 +8077,8 @@
     #
 
     playwright.test.project[playwright-js-tomcat90-mysql57][publications]=\
-        change-tracking-web
+        change-tracking-web-1,\
+        change-tracking-web-2
 
     test.batch.class.names.includes[publications]=\
         **/apps/**/src/testIntegration/**/mvcc/**/*Test.java,\
